chore: changing exports
